# openapi.yaml
openapi: 3.0.3
info:
  title: Lab API
  version: 1.0.0
  description: >-
    REST API for laboratory management, document workflows, and OnlyOffice
    integration. JWT bearer authentication with scope-based access control.
servers:
  - url: http://localhost:8000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        scopes:
          type: array
          items:
            type: string
        expires_in:
          type: integer
    Doc:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        path:
          type: string
        description:
          type: string
          nullable: true
        owner_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Lab:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Sample:
      type: object
      properties:
        id:
          type: integer
        lab_id:
          type: integer
        code:
          type: string
        status:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
paths:
  /healthz:
    get:
      summary: Liveness probe
      responses:
        '200':
          description: Service healthy
  /health:
    get:
      summary: Database-backed health probe
      responses:
        '200':
          description: Service healthy
        '500':
          description: Dependency error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth:
    post:
      summary: Exchange API key for JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [api_key]
              properties:
                api_key:
                  type: string
      responses:
        '200':
          description: Token bundle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid API key
  /auth/login:
    post:
      summary: Authenticate with username/email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Token bundle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      summary: Refresh access token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Refreshed token bundle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /auth/change_password:
    post:
      summary: Change password for the logged-in user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: Password updated
        '401':
          description: Invalid credentials
  /auth/request_password_reset:
    post:
      summary: Request a password reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Reset token issued (for demo/testing)
  /auth/perform_password_reset:
    post:
      summary: Complete password reset using token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Password reset
        '400':
          description: Invalid token
  /auth/register:
    post:
      summary: Register a user via invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password, invite_code]
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                invite_code:
                  type: string
      responses:
        '201':
          description: User created and authenticated
        '400':
          description: Invalid invite
  /auth/create_invite:
    post:
      summary: Create invitation codes (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  nullable: true
                expires_in_hours:
                  type: integer
                  default: 72
                max_uses:
                  type: integer
                  default: 1
      responses:
        '201':
          description: Invite generated
        '403':
          description: Missing admin scope
  /api/v1/docs:
    get:
      summary: List documents
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated docs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Doc'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      size:
                        type: integer
        '403':
          description: Missing doc scope
  /api/v1/docs/{doc_id}:
    get:
      summary: Fetch a single document
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: doc_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doc'
        '404':
          description: Document not found
  /api/v1/docs/{doc_id}/edit:
    get:
      summary: Generate OnlyOffice editor config for a document
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: doc_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OnlyOffice editor payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    type: object
                  documentUrl:
                    type: string
                    format: uri
        '404':
          description: Document not found
  /api/v1/docs/{doc_id}/callback:
    post:
      summary: Receive OnlyOffice callback payloads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Callback acknowledged
  /files/{path}:
    get:
      summary: Serve document files for OnlyOffice
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
  /api/v1/meta:
    get:
      summary: List metadata for whitelisted tables
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Table metadata
  /api/v1/meta/{table}:
    get:
      summary: Inspect a specific table's columns
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Column metadata
        '404':
          description: Table unknown
  /api/v1/table/{table}:
    get:
      summary: List rows for a table
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: size
          schema:
            type: integer
      responses:
        '200':
          description: Records
    post:
      summary: Insert a record into a table
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Record created
        '400':
          description: Validation error
  /api/v1/table/{table}/{pk}:
    put:
      summary: Update a record
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: pk
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Record updated
        '404':
          description: Not found
    delete:
      summary: Delete a record
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: pk
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Record removed
  /api/v1/labs:
    get:
      summary: List labs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lab collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lab'
    post:
      summary: Create a lab
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                location:
                  type: string
      responses:
        '201':
          description: Lab created
  /api/v1/labs/{lab_id}:
    get:
      summary: Retrieve a lab
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: lab_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lab detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Lab'
    put:
      summary: Update a lab
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: lab_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Lab updated
    delete:
      summary: Delete a lab
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: lab_id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Lab deleted
  /api/v1/samples:
    get:
      summary: List samples
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sample collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Sample'
    post:
      summary: Create a sample
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lab_id, code]
              properties:
                lab_id:
                  type: integer
                code:
                  type: string
                status:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Sample created
  /api/v1/samples/{sample_id}:
    get:
      summary: Retrieve a sample
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: sample_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sample detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Sample'
    put:
      summary: Update a sample
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: sample_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Sample updated
    delete:
      summary: Delete a sample
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: sample_id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Sample deleted
security:
  - bearerAuth: []